requirements.txt
yfinance>=0.2.18
requests>=2.31.0
pandas>=2.0.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langgraph>=0.1.0
python-dotenv>=1.0.0
openai>=1.0.0